refactor(text-to-image): moves SharkUI SDXL client to "TextToImage/client/SDXL"
