+++ src/features/TextToImage/client/SDXL/conversions/GenerateFromTextResponse.ts
