+++ src/features/TextToImage/client/SDXL/conversions/StabilityAI_Client_Image.ts
