+++ src/features/TextToImage/client/exports.ts
@@ -1,3 +1,3 @@
 export {
   default as SDXL,
-} from '../models/SDXL/client';
+} from './SDXL';
